+++ src/vivarium_public_health/testing/mock_artifact.py
