+++ tests/population/test_base_population.py
@@ -9,8 +9,8 @@ from vivarium.testing_utilities import get_randomness
 
 import vivarium_public_health.population.base_population as bp
 import vivarium_public_health.population.data_transformations as dt
+from tests.test_utilities import make_uniform_pop_data
 from vivarium_public_health import utilities
-from vivarium_public_health.testing.utils import make_uniform_pop_data
 
 
 def test_select_sub_population_data():
